Extract createTabsSlice from projectStore (P5 store slice)
